refactor, delete duplicated code! - move stuff to functions, move functions out to a central place - capture globals in a single place
